backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn==0.24.0
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
APScheduler==3.10.4